Move the hydropower test to the test_river module as the reservoir node is in the river module
